Refactor graph tools registration into a separate module
